requirements.txt
streamlit>=1.35.0
supabase
python-dotenv
plotly
fpdf2
Pillow
streamlit-lottie
calplot
pytz
requests
pandas
matplotlib